--- a/compiler/src/Language/Mimsa/Actions/Interpret.hs
+++ b/compiler/src/Language/Mimsa/Actions/Interpret.hs
@@ -38,6 +38,9 @@ interpreter se = do
     Just re -> pure (bimap fst edAnnotation re)
     _ -> throwError (StoreErr (CouldNotFindStoreExpression (getStoreExpressionHash se)))
 
+squashify :: (Ord e) => Map e (Map e a) -> Map e a
+squashify = mconcat . M.elems
+
 -- Interpret a group of StoreExpressions
 -- This means each sub dep is only interpreted once
 interpretAll ::
@@ -45,10 +48,16 @@ interpretAll ::
   Actions.ActionM (Map ExprHash (InterpretExpr Name Annotation))
 interpretAll inputStoreExpressions = do
   let action depMap se = do
+        -- get us out of this Map of Maps situation
+        let flatDeps = squashify depMap
         -- tag each `var` with it's location if it is an import
         let withImports = addEmptyStackFrames (convertImports se)
         -- interpret se
-        liftEither (first InterpreterErr (interpret depMap withImports))
+        interpreted <- liftEither (first InterpreterErr (interpret flatDeps withImports))
+        -- we need to accumulate all deps
+        -- as we go, so pass them up too
+        let allDeps = flatDeps <> M.singleton (getStoreExpressionHash se) interpreted
+        pure allDeps
 
   -- create initial state for builder
   -- we tag each StoreExpression we've found with the deps it needs
@@ -68,4 +77,5 @@ interpretAll inputStoreExpressions = do
                 <$> inputStoreExpressions,
             Build.stOutputs = mempty -- we use caches here if we wanted
           }
-  Build.stOutputs <$> Build.doJobs action state
+  splat <- Build.stOutputs <$> Build.doJobs action state
+  pure (squashify splat)