Interpret all dependencies of an expression before interpreting the expression itself.

Before:
```
benchmarking evaluate/evaluate big looping thing
time                 145.4 ms   (140.4 ms .. 150.2 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 144.7 ms   (143.0 ms .. 147.6 ms)
std dev              3.087 ms   (1.651 ms .. 4.432 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking evaluate/evaluate parsing
time                 108.1 ms   (104.9 ms .. 110.0 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 109.9 ms   (108.4 ms .. 112.9 ms)
std dev              3.344 ms   (1.594 ms .. 5.177 ms)

benchmarking evaluate/evaluate parsing 2
time                 29.80 s    (29.51 s .. 29.97 s)
                     1.000 R²   (NaN R² .. 1.000 R²)
mean                 29.72 s    (29.63 s .. 29.77 s)
std dev              92.37 ms   (37.94 ms .. 127.6 ms)
variance introduced by outliers: 19% (moderately inflated)
```

After:

```
benchmarking evaluate/evaluate big looping thing
time                 145.7 ms   (142.6 ms .. 149.4 ms)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 146.7 ms   (144.2 ms .. 148.8 ms)
std dev              3.370 ms   (2.261 ms .. 5.449 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking evaluate/evaluate parsing
time                 102.7 ms   (99.77 ms .. 105.4 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 109.0 ms   (106.7 ms .. 112.7 ms)
std dev              4.731 ms   (2.242 ms .. 7.810 ms)

benchmarking evaluate/evaluate parsing 2
time                 27.53 s    (24.46 s .. 29.15 s)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 28.02 s    (27.59 s .. 28.34 s)
std dev              428.5 ms   (205.0 ms .. 583.7 ms)
variance introduced by outliers: 19% (moderately inflated)
```